Move mappers to own package
