+++ src/ext/java/org/opentripplanner/ext/siri/SiriAlertsUpdateHandler.java
@@ -7,6 +7,8 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import org.opentripplanner.ext.siri.mapper.AffectsMapper;
+import org.opentripplanner.ext.siri.mapper.SiriSeverityMapper;
 import org.opentripplanner.framework.i18n.I18NString;
 import org.opentripplanner.framework.i18n.NonLocalizedString;
 import org.opentripplanner.framework.i18n.TranslatedString;
